consolidate workflow
