--- javascript/workflow/src/engine/index.ts
@@ -1,9 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from '../resources/Workflow.js'
-export * from './createWorkflow.js'
-export * from './readStep.js'
-export * from './types.js'
-export * from './validator.js'
