+++ javascript/workflow/src/index.ts
@@ -3,7 +3,6 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './dataflow/index.js'
-export * from './engine/index.js'
 export * from './primitives.js'
 export * from './resources/index.js'
 export * from './util/index.js'
