+++ javascript/workflow/src/resources/Workflow/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright (c) Microsoft. All rights reserved.
+ * Licensed under the MIT license. See LICENSE file in the project.
+ */
+export * from './Workflow.js'
+export * from './Workflow.types.js'
