+++ javascript/workflow/src/resources/index.ts
@@ -11,4 +11,4 @@ export * from './Observed.js'
 export * from './ParserOptions.js'
 export * from './Resource.js'
 export * from './types.js'
-export * from './Workflow.js'
+export * from './Workflow/index.js'
